Only create swap parameters once to avoid mismatch in expiries
