+++ api_tests/tests/dry/lightning_routes.ts
@@ -8,11 +8,13 @@ import {
     defaultHalightLightningBitcoinHerc20EthereumErc20,
     defaultHanEthereumEtherHalightLightningBitcoin,
     defaultHerc20EthereumErc20HalightLightningBitcoin,
-} from "../../src/actors/defaults";
+} from "../../src/actors/swap_factory";
 
 // ******************************************** //
 // Lightning routes                               //
 // ******************************************** //
+
+// TODO: All these tests should use the SwapFactory, these `default` functions are broken as soon as we use both actors because the expiries are created with Date.now() and hence different for both actors
 describe("Lightning routes tests", () => {
     it(
         "lightning-routes-post-eth-lnbtc-return-201",
