+++ api_tests/tests/e2e/bitcoin_ethereum.ts
@@ -7,6 +7,7 @@ import { twoActorTest } from "../../src/actor_test";
 import { AssetKind } from "../../src/asset";
 import { HarnessGlobal, sleep } from "../../src/utils";
 import { LedgerKind } from "../../src/ledgers/ledger";
+import SwapFactory from "../../src/actors/swap_factory";
 
 declare var global: HarnessGlobal;
 
@@ -72,11 +73,13 @@ describe("E2E: Ethereum/ether - Lightning/bitcoin", () => {
     it(
         "han-ethereum-ether-halight-lightning-bitcoin-alice-redeems-bob-redeems",
         twoActorTest(async ({ alice, bob }) => {
+            const [aliceBody, bobBody] = await SwapFactory.newSwap(alice, bob);
+
             // make sure bob knows about the swap first
-            await bob.createSwap();
+            await bob.createSwap(bobBody);
             await sleep(500);
 
-            await alice.createSwap();
+            await alice.createSwap(aliceBody);
 
             await alice.init();
